-Modified Validations for onsubmit removing unnecessary validations -set spinner clockwise
