+++ src/pages/Designer.js
@@ -49,6 +49,7 @@ import DownloadButton from '../components/DownloadButton';
 import ContextMenu from '../components/ContextMenu';
 import CustomNodeModal from '../components/Modal/CustomNodeModal';
 import Generating from '../components/Generating';
+import { checkDisabled } from '../utils/submitButtonValidation';
 
 let serviceId = 1;
 let gatewayId = 1;
@@ -366,12 +367,6 @@ const Designer = ({ update, viewMode = false, sharedMetadata = undefined }) => {
                             setServiceDiscoveryCount(0);
                             setIsServiceDiscovery(false);
                             for (let key in updatedEdges) {
-                                let Edge = updatedEdges[key];
-                                if (Edge?.data?.framework === 'rest-api') {
-                                    delete Edge?.data?.type;
-                                    delete Edge?.data?.framework;
-                                    delete Edge?.label;
-                                }
                                 if (key.split('-')[1] == 'serviceDiscoveryType') {
                                     delete updatedEdges[key];
                                 }
@@ -1797,94 +1792,6 @@ const Designer = ({ update, viewMode = false, sharedMetadata = undefined }) => {
 
     const onPaneClick = useCallback(() => setMenu(false), [setMenu]);
 
-    const checkNodeExists = Object.values(nodes).some(
-        node => node.id.startsWith('Service') || node.id.startsWith('Gateway') || node.id.startsWith('UI'),
-    );
-
-    const projectNameCheck = !/^[a-zA-Z](?:[a-zA-Z0-9_ -]*[a-zA-Z0-9])? *$/.test(projectName);
-
-    const isDatabaseConnectedAndFilled = () => {
-        let dbConnected = false;
-        let dbFilled = false;
-
-        for (const key in nodes) {
-            let databaseCheck = nodes[key];
-            if (databaseCheck?.id?.startsWith('Database')) {
-                if (!databaseCheck?.data?.isConnected) {
-                    dbConnected = true;
-                }
-                if (!databaseCheck?.data?.databasePort) {
-                    dbFilled = true;
-                }
-            }
-            if (dbConnected && dbFilled) {
-                break;
-            }
-        }
-
-        return { isConnected: dbConnected, isFilled: dbFilled };
-    };
-
-    const checkDisabled = () => {
-        const databaseStatus = isDatabaseConnectedAndFilled();
-
-        if (!checkNodeExists) {
-            return { isValid: false, message: 'Drag and drop atleast one Application to generate the code' };
-        }
-
-        if (!projectName || projectNameCheck) {
-            return { isValid: false, message: 'Architecture name should be valid.' };
-        }
-
-        if (projectName === '') {
-            return { isValid: false, message: 'Architecture name is empty.' };
-        }
-
-        if (isEmptyUiSubmit) {
-            return { isValid: false, message: 'UI is not Configured. Click on the highlighted UI Node to Configure it.' };
-        }
-
-        if (isEmptyServiceSubmit) {
-            return { isValid: false, message: 'Service is not Configured. Click on the highlighted Service node to Configure it.' };
-        }
-
-        if (isEmptyGatewaySubmit) {
-            return { isValid: false, message: 'Gateway is not Configured. Click on the highlighted Gateway node to Configure it.' };
-        }
-
-        if (databaseStatus.isConnected) {
-            return { isValid: false, message: 'Create an edge connecting the node to Database to enable the integration.' };
-        }
-
-        if (databaseStatus.isFilled) {
-            return { isValid: false, message: 'Database is not Configured. Click on the highlighted Database node to Configure it.' };
-        }
-
-        return { isValid: true, message: 'Validation successful. Proceed to generate the application.' };
-    };
-
-    const checkEdge = () => {
-        let updatedEdges = { ...edges };
-        let updatedNodes = { ...nodes };
-        if (Object.keys(updatedNodes).length !== 0) {
-            for (const key in updatedNodes) {
-                let databaseCheck = updatedNodes[key];
-                if (databaseCheck?.id?.startsWith('Database') && !databaseCheck?.data?.isConnected) {
-                    return true;
-                }
-            }
-        }
-        if (Object.keys(updatedEdges).length !== 0) {
-            for (const key in updatedEdges) {
-                let edgeCheck = updatedEdges[key];
-                if (edgeCheck?.target?.startsWith('Service') && !edgeCheck?.data?.framework) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    };
-
     const handleSave = async () => {
         if (projectName && projectName !== 'clear#canvas') {
             if (!keycloak.authenticated) {
@@ -1899,7 +1806,7 @@ const Designer = ({ update, viewMode = false, sharedMetadata = undefined }) => {
     };
 
     const handleSubmit = () => {
-        const { isValid, message } = checkDisabled();
+        const { isValid, message } = checkDisabled(projectName, isEmptyUiSubmit, isEmptyServiceSubmit, isEmptyGatewaySubmit, nodes, edges);
         const errorMessage = message || 'Validation failed';
         toast.close(toastIdRef.current);
         toastIdRef.current = toast({
@@ -1909,7 +1816,6 @@ const Designer = ({ update, viewMode = false, sharedMetadata = undefined }) => {
             variant: 'left-accent',
             isClosable: true,
         });
-        if (checkEdge()) return;
         if (!isValid) return;
         saveData(false);
     };
